+++ test-at/src/test/scala/com/stratio/sparkta/testat/embedded/ElasticsearchEmbeddedServer.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.stratio.sparkta.testat.embeddedes
+package com.stratio.sparkta.testat.embedded
 
 import java.io.File
